style: 首页的格式统一

+++ src/views/acl/user/index.vue
@@ -1,27 +1,16 @@
 <template>
   <div>
     <el-card style="margin: 10px 0">
-      <el-from
-        :inline="true"
-        style="
+      <el-from :inline="true" style="
           display: flex;
           justify-content: space-between;
           align-items: center;
-        "
-      >
+        ">
         <el-form-item label="用户名">
-          <el-input
-            placeholder="请输入搜索用户名"
-            style="width: 200px"
-            v-model="serachvalue"
-          ></el-input>
+          <el-input placeholder="请输入搜索用户名" style="width: 200px" v-model="serachvalue"></el-input>
         </el-form-item>
         <el-form-item>
-          <el-button
-            type="success"
-            @click="searchuser"
-            :disabled="!serachvalue"
-          >
+          <el-button type="success" @click="searchuser" :disabled="!serachvalue">
             搜索
           </el-button>
           <el-button type="success" @click="resetuser">重置</el-button>
@@ -30,17 +19,8 @@
     </el-card>
     <el-card>
       <el-button type="success" @click="adduser">添加用户</el-button>
-      <el-popconfirm
-        title="你确认要删除选中的用户吗？"
-        confirmButtonText="删除"
-        cancelButtonText="取消"
-        confirmButtonType="success"
-        cancelButtonType="text"
-        icon="Delete"
-        iconColor="#f90"
-        @confirm="deletesomeuser"
-        v-has="`btn.User.add`"
-      >
+      <el-popconfirm title="你确认要删除选中的用户吗？" confirmButtonText="删除" cancelButtonText="取消" confirmButtonType="success"
+        cancelButtonType="text" icon="Delete" iconColor="#f90" @confirm="deletesomeuser" v-has="`btn.User.add`">
         <template #reference>
           <el-button type="success" v-has="`btn.User.remove`">
             批量删除
@@ -48,14 +28,7 @@
         </template>
       </el-popconfirm>
 
-      <el-table
-        border
-        stripe
-        @selection-change=""
-        style="margin: 10px 0"
-        :data="adminuserinfo"
-        ref="reftable"
-      >
+      <el-table border stripe @selection-change="" style="margin: 10px 0" :data="adminuserinfo" ref="reftable">
         <el-table-column type="selection" width="55" />
         <el-table-column type="index" width="50" label="#" />
         <el-table-column label="ID" width="50" prop="userid"></el-table-column>
@@ -64,38 +37,16 @@
         <el-table-column label="用户角色" prop="role"></el-table-column>
         <el-table-column label="操作">
           <template #="{ row }">
-            <el-button
-              type="success"
-              icon="User"
-              @click="updateuserrole(row)"
-              v-has="`btn.User.assgin`"
-            >
+            <el-button type="success" icon="User" @click="updateuserrole(row)" v-has="`btn.User.assgin`">
               分配角色
             </el-button>
-            <el-button
-              type="success"
-              icon="Edit"
-              @click="edituser(row)"
-              v-has="`btn.User.update`"
-            >
+            <el-button type="success" icon="Edit" @click="edituser(row)" v-has="`btn.User.update`">
               编辑
             </el-button>
-            <el-popconfirm
-              title="你确认要删除这个用户吗？"
-              confirmButtonText="删除"
-              cancelButtonText="取消"
-              confirmButtonType="success"
-              cancelButtonType="text"
-              icon="Delete"
-              iconColor="#f90"
-              @confirm="deleteuser(row)"
-            >
+            <el-popconfirm title="你确认要删除这个用户吗？" confirmButtonText="删除" cancelButtonText="取消" confirmButtonType="success"
+              cancelButtonType="text" icon="Delete" iconColor="#f90" @confirm="deleteuser(row)">
               <template #reference>
-                <el-button
-                  type="success"
-                  icon="Delete"
-                  v-has="`btn.User.remove`"
-                >
+                <el-button type="success" icon="Delete" v-has="`btn.User.remove`">
                   删除
                 </el-button>
               </template>
@@ -103,45 +54,22 @@
           </template>
         </el-table-column>
       </el-table>
-      <el-pagination
-        style="margin-top: 10px"
-        @size-change="sizeChange"
-        @current-change="getadminuser"
-        v-model:currentPage="pageNo"
-        :page-sizes="[3, 6, 9]"
-        :page-size="pageSize"
-        layout=" prev, pager, next, jumper,->, sizes,total"
-        :total="total"
-      >
-        :pager-count="7">
+      <el-pagination style="margin-top: 10px" @size-change="sizeChange" @current-change="getadminuser"
+        v-model:currentPage="pageNo" :page-sizes="[3, 6, 9]" :page-size="pageSize"
+        layout=" prev, pager, next, jumper,->, sizes,total" :total="total" v-if="!(adminuserinfo.length>pageSize)">
       </el-pagination>
     </el-card>
     <el-drawer title="添加用户" v-model="drawer" size="30%">
       <template #default>
         <el-form ref="formRef" :model="draweruserinfo" :rules="rules">
           <el-form-item label="用户名" label-width="100px" prop="username">
-            <el-input
-              placeholder="请输入用户名"
-              v-model="draweruserinfo.username"
-            ></el-input>
+            <el-input placeholder="请输入用户名" v-model="draweruserinfo.username"></el-input>
           </el-form-item>
-          <el-form-item
-            label="密码"
-            label-width="100px"
-            prop="password"
-            v-if="draweruserinfo.userid == 0 ? true : false"
-          >
-            <el-input
-              placeholder="请输入密码"
-              type="password"
-              v-model="draweruserinfo.password"
-            ></el-input>
+          <el-form-item label="密码" label-width="100px" prop="password" v-if="draweruserinfo.userid == 0 ? true : false">
+            <el-input placeholder="请输入密码" type="password" v-model="draweruserinfo.password"></el-input>
           </el-form-item>
           <el-form-item label="用户昵称" label-width="100px" prop="name">
-            <el-input
-              placeholder="请输入昵称"
-              v-model="draweruserinfo.name"
-            ></el-input>
+            <el-input placeholder="请输入昵称" v-model="draweruserinfo.name"></el-input>
           </el-form-item>
         </el-form>
       </template>
@@ -151,22 +79,11 @@
       </template>
     </el-drawer>
     <el-drawer title="分配角色" v-model="drawer1" size="30%">
-      <el-checkbox
-        v-model="checkAll"
-        :indeterminate="isIndeterminate"
-        @change="handleCheckAllChange"
-      >
+      <el-checkbox v-model="checkAll" :indeterminate="isIndeterminate" @change="handleCheckAllChange">
         全选
       </el-checkbox>
-      <el-checkbox-group
-        v-model="rolechecked"
-        @change="handleCheckedRoleChange"
-      >
-        <el-checkbox
-          v-for="item in role"
-          :key="item.id"
-          :label="item.rolename"
-        ></el-checkbox>
+      <el-checkbox-group v-model="rolechecked" @change="handleCheckedRoleChange">
+        <el-checkbox v-for="item in role" :key="item.id" :label="item.rolename"></el-checkbox>
       </el-checkbox-group>
       <template #footer>
         <el-button type="info" @click="drawer1 = false">取消</el-button>
@@ -202,6 +119,7 @@ let draweruserinfo = reactive({
 })
 const reftable = ref()
 let userid = ref(0)
+
 onMounted(() => {
   getadminuser()
 })
@@ -371,8 +289,6 @@ const setuserrole = async () => {
 }
 const searchuser = async () => {
   const result = await SearchUser(
-    pageNo.value,
-    pageSize.value,
     serachvalue.value,
   )
   if (result.code == 200) {
@@ -389,11 +305,19 @@ const deletesomeuser = async () => {
   useridattr = useridattr.map((item: any) => {
     return item.userid
   })
-  const result = await DeleteUser(useridattr)
-  if (result.code == 200) {
-    ElMessage({ type: 'success', message: '删除成功' })
-    window.location.reload()
+  if (useridattr.length > 0) {
+    const result = await DeleteUser(useridattr)
+    if (result.code == 200) {
+      ElMessage({ type: 'success', message: '删除成功' })
+      window.location.reload()
+    }
+  } else {
+    ElMessage({
+      type: 'error',
+      message: '请选择你需要删除的用户'
+    })
   }
+
 }
 const deleteuser = async (row: any) => {
   const useridattr = [row.userid]
@@ -406,6 +330,7 @@ const deleteuser = async (row: any) => {
     window.location.reload()
   }
 }
+
 </script>
 
 <style scoped></style>
